V2.1 development
All printers and variants now have been updated to 2.1

The 2.1 configuration and macro files are much more streamlined. All the files are now identical, with the exception of the bed.g (different between the Duet and Duex enabled printers) and the config.g (different for every printer).

All the macro files are now identical for all printer and printer variants